Do not depend on isl for building gcc
Build gcc without isl, because the version gcc 5 depends on is too old
to be available as a binary package in recent distributions.

Note that in doing so we lose a potential loop optimization pass. So
this library should be reused eventually, but directly from a source
tarball and compiled statically into gdb.

Resolves: #21