PyGame: disable audio and video init to allow headless environments
When running headless (e.g. over ssh or in CI), PyGame fails to
initialize because SDL cannot access any audio hardware. However
we only use it as a convenient and portable library for image
loading and manipulation, we don't depend on its audio or video
capabilities.

We can't selectively disable PyGame's audio or video subsystem,
but we can force SDL2 to use a dummy backend, which is enough
to allow paltool and tiletool to run under headless environments.