Do not install into ngdevkit-toolchain's prefix
Various runtime files (e.g. crt0) were installed in gcc's install path.
This extrapolated install location is not valid when every git module
is installed in a separate prefix, as is it the case for instance with
brew in macOS.

Fix the install to split the install paths if needed, and stick to
the original merged path otherwise.