mingw: force configure to recognize cross-compilation
When running a Linux environment under WSL2 (win10), w64 binaries
are transparently executed as if they were native Linux binaries.
This fools the autoconf scripts which don't recognize that we're
cross-compiling for win10.

To prevent all sorts of bad side effect on compilation, force
autoconf scripts to enable cross-compilation flags when mingw is
used.