Reimplementation of nullsound sound driver
Breaking change.

nullsound is now a z80 library that can be linked by users. The
sound commands are now implemented by the user, which just calls
a series of functions from the nullsound library.

nullsound internally is now much simpler. It provides simple actions
for trigerring immediate sounds; it manages a state tracker for every
part of the YM2610.

This rebuild prepares nullsound for managing interrupts and playback
of simple sound fx or music via VDM.