Support music playback via nullsound streams
This is a initial commit to play a music encoded in the nullsound stream
format (NSS), which is as a series of opcodes that can configure various
parts of the YM2610, wait for timers, and control music loop.

The NSS format is still considered experimental and will change in
future commits.