NSS: add ADPCM-A and FM context opcodes
Currently ADPCM-A and FM opcodes (instrument, note on...) are
two bytes long and they include the action's target channel.

Optimize the space used by defining a 10 new CTX opcodes.
A CTX is a 1 byte opcode that defines the scope of the next
ADPCM-A or FM action.

ADPCM-A and FM have their own context of action. Each time
a ADPCM-A or FM opcode is evaluated, the according context
is incremented by 1, so that the scope of the next opcode
will automatically be the next ADPCM-A or FM channel.

This approach allows efficient NSS stream storage in ROM, and
will be reused when implementing the next NSS opcodes
like e.g. volume, detune...

The original ADPCM-A and FM opcodes are now deprecated and
will be replaced eventually.