ngdevkit 0.4: new tools for VROM asset management
vromtool.py automatically packs ADPCM samples into one or more
VROMs, and generates ASM defines to export the samples positions
in VROM to the z80 sound driver.

furtool.py extracts the FM and ADPCM instruments define in a
of a Furnace module, and generate instrument data ready to be
consumed by nullsound NSS stream player. It also extract the
ADPCM samples from the Furnace module in a sample map that
can be consumed by vromtool.py to generate VROMs.

Bump ngdevkit version to 0.4.