nullsound: fix YM2610 reset before stream playback
FM channels can't be stopped immediately, as their operators can
only be configured to release their note.

Implement the reset by configuring the fastest release possible
and stopping the operator in sequence. This is sufficient to
remove any unpleasant "pop" sound that was appearing while
stopping a NSS stream and playing another one immediately.

Also update the snd_stream_stop to call ym2610_reset so that
all channels are effectively stopped when this function is called.