nullsound: Do not use ar for packing nullsound files
the nullsound library is a collection of files packaged into an
ar archive to ease linking of music ROMs.
However it looks like passing the library as a linker input as if
it was a simple .rel object is not supported by SDCC, and it can
break sdld parsing. This is now happening on the MSYS2 environment.

In order to keep nullsound as a standalone file and fix sdld linking,
make nullsound.lib a simple concatenation of all .rel objects. This
is correctly parsed by sdld, on all environments (Linux, macOS, MSYS2).