CI: tweak MSYS2 package download to avoid timeout
The ngdevkit packages are hosted on github pages, and the throughput
between this hosting location and github CI runner seem to fluctuate,
which sometimes causes pacman to timeout when installing packages,
and prevent the build of nightly packages.

Disable timeout checks in pacman to give CI runner a chance to install
MSYS2 packages even when it's temporarily slow.