nullsound: rework SSG macros and how they are applied
Two major updates to how SSG macro work in nullsound:

1.
The logic of update_ssg_macro used run macro steps for each
SSG channel and then load the mirrored state of all channels.
This is not ideal as it cannot detect when a macro is finished,
and thus it keep uploading mirrored state into the YM2610 at
every step even when not necessary.
Now the algorithm group update and load together, so it won't
load data into the YM2610 when a macro is finished.

2.
The SSG register common to all channels must not be updated at
every tick. For instance, when doing so with the envelope channel,
this resets the envelope shape on every note, which is not the
behaviour implemented by Furnace.
Split macros into two subparts: one is only called once when the
macro is set up for a SSG channel. The other parts contains the
remaining data that must be run on every tick while a note is
playing.