nullsound: check validity of 8bit pointer addition
In order to optimize speed and size of Z80 code, one can do
8bit addition on 16bit pointers, assuming the result of the addition
will not change the MSB of the pointer.
This works as long as nullsound data are linked appropriately in
the Z80 address space by using the right .bndry statements.

To verify that nullsound binaries have the expected boundary
properties, add a validating script that checks that data surrounded
by labels _state_XXX_start to _state_XXX_end are always allocated
by sdld to addresses with the same MSB.