New compact representation for NSS opcodes
nullsound now supports playing multiple NSS streams concurrently, as
well as a new, more compact representation of the music's channels.

This is achieved by a couple of things:
  . Adding new NSS opcodes such as call/ret, which allows storing a
    pattern's data only once if this pattern is repeated in the music
  . Saving channels as individual NSS streams the deduplication with
    call/ret, as well as future opcode compaction strategies.

By default, the old "inline" NSS stream is still generated by nsstool,
the switch to compact format or the deprecation of the inline format
will be assessed later.