nullsound: make YM2610 access reentrant to fix music playback bug
The interrupt handler needs to write to the YM2610 to rearm timer
when an interrupt is triggered. This can happen while the sound
driver is configuring the YM2610 to write to a specific register.

To avoid any register misconfiguration, make writes to port A
reentrant, and restore the register context when exiting from
the interrupt handler