soundtool.py: sound asset management for nullsound
soundtool.py is a new tool that automates the generation of sound maps
and sound commands for m68k and z80.

soundtool.py supports both Furnace (music) and ADPCM (sfx) in sound maps,
and is able to generate sound command header for the m68k code, as well
as generate the boilerplate z80 code to trigger music or sound fx from
the z80.