nullsound: global music volume support
Add functions to control the global output volume for a playing
NSS stream.
Add volume fade out function for both music and SFX.
The fade out support currently doesn't support directly modifying
OPx volumes in furnace. This will be implemented in a subsequent commit.

This commit introduces a change in how nsstool generates stream
data, and nullsound now expects this new format, so it is not
backward-compatible with previously generated NSS data.
The instrument data generated by furtool also slightly differs and
is not backward-compatible.