nullsound: fix race when handling NMI and INT concurrently
When a Timer interrupts the current sound driver, the interrupt
handler has to save current registers. During this execution,
new interrupts are masked, but a NMI can still occur and interrupt
the interrupt handler.

To avoid losing the state of the original sound driver, one cannot
use shadow registers for both Timer and NMI handlers. Instead,
make the Timer interrupt handler save register on the stack.

Closes #115